requirements.txt
streamlit==1.31.0
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
chardet==5.2.0
xlrd==2.0.1
msoffcrypto-tool==5.0.1
olefile==0.47
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
